Add more abstraction to models
